% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_grob.R
\name{convert_to_grob}
\alias{convert_to_grob}
\title{The function used within grob_col which takes in the object the 
user wants to grob, and converts it to a grob.}
\usage{
convert_to_grob(x, height, width, more_args = list())
}
\arguments{
\item{x}{The object which needs to be converted to a grob. Must be either:
A data.frame/martrix, the name of a .png image, a character string, a 
ggplot, NA (for an empty grob), or already a grob.}

\item{height}{The numeric height in mm of the desired grob.}

\item{width}{The numeric width in mm of the desired grob.}

\item{more_args}{A list with any extra aesthetic preferences for the desired grobs.}
}
\value{
A grob of x with aesthetics based on the more_args param.
}
\description{
The function used within grob_col which takes in the object the 
user wants to grob, and converts it to a grob.
}
